% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as_num}
\alias{as_num}
\title{Testa se o valor eh numerico. Se nao for, converte.}
\usage{
as_num(i, check = TRUE, fix_na = TRUE)
}
\arguments{
\item{i}{vetor qualquer}

\item{check}{TRUE checa a estrutura do dado}

\item{fix_na}{vetor de strings a serem removidas, caso existam.}
}
\value{
vetor numerico tratado a partir de i
}
\description{
Verifica se o conteudo do vetor passado e de numeros. Quando check = TRUE, a estrutura do dado
e testada via regex e se o padrao for detectado, o vetor e definido como numerico.
}
