% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{converte_to_json}
\alias{converte_to_json}
\title{Salvar data frame como json}
\usage{
converte_to_json(
  x,
  file,
  df_type = "rows",
  raw_type = "mongo",
  POSIXt = "mongo",
  factor = "string",
  trace = FALSE
)
}
\arguments{
\item{x}{data.frame}

\item{file}{destino do arquivo, incluindo extensao .json}

\item{df_type}{salvar no formato linha, coluna ou valor}

\item{raw_type}{tipo dos dados brutos de saida}

\item{POSIXt}{formato de datetime}

\item{factor}{tratar fator como string?}

\item{trace}{se TRUE, exibe estatisticas do arquivo salvo}
}
\description{
Recebe um dataframe transforma em json e salva no caminho especiicado
}
