% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{count_numeric_digits}
\alias{count_numeric_digits}
\title{Conta total de digitos antes ou sepois da virgula}
\usage{
count_numeric_digits(x, depois = TRUE)
}
\arguments{
\item{x}{Vetor numerico}

\item{depois}{Se TRUE, retorna total de digitos depois da virgula, caso contrario retorna total de digitos da parte inteira.}
}
\description{
Conta total de digitos antes ou sepois da virgula
}
