% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\docType{data}
\name{dm_local}
\alias{dm_local}
\title{Dados do censo ensino superior INEP 2017 (Local da oferta)}
\description{
Este conjunto de dados faz parte da base de dados do INEP para o censo
brasileiro do ensino superior.
Nesta parte temos os dados dos locais de oferta do curso das
Instituicoes de Ensino Superior mapeados pelo censo naquele ano.
Trata-se de uma amostras das variaveis, porem o total de observacoes
esta completo.
Para obter dados de outros anos consulte o link do site do INEP ou
a funcao \code{\link{rnp_get_inep_censo}}
}
\examples{
require(rnpextras)
str(dm_local)
}
\references{
\url{https://www.gov.br/inep/pt-br/acesso-a-informacao/dados-abertos/microdados/censo-da-educacao-superior}
}
\author{
INEP
}
\keyword{dm_local}
