% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gerador_de_senha}
\alias{gerador_de_senha}
\title{Gerador de senha forte (ou absurdas)}
\usage{
gerador_de_senha(nsenhas = 1, tamanho = 16, seed = NULL, md5 = TRUE)
}
\arguments{
\item{nsenhas}{total de senhas. Se colocar 20 saem vinte senhas}

\item{tamanho}{tamanho da senha em caracteres}

\item{seed}{numerico para pode repetir as senhas geradas antes. Se nao for informado usara smotragem por decimal uniforme entre 0 e 1}

\item{md5}{Padrao TRUE, Devolve a senha tambem em MD5. Se FALSE devolve apenas as senhas}
}
\value{
Vetor ou data.frame com as senhas
}
\description{
Gerador de senha forte (ou absurdas)
}
\details{
Dado o total de senhas e o tamanho delas, o programa calcula aleatoreamente
as senhas com as seguintes regras:
1. Tem letras maiusculas
2. Tem letras minusculas
3. Tem numeros de 0 a 9
4. Tem caracteres especiais ex:  "#","$","*",".","!","?"
}
\examples{
# 10 senhas tamanho 8 com senha de seed de rastreamento 1234 mais MD5
gerador_de_senha(10, 8, 1234, TRUE)
# O mesmo sem seed
gerador_de_senha(10, 8)
# Agora sem seed nem md5 e com tamanho 12
gerador_de_senha(10, 12, NULL, FALSE)
}
