% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_cnpj}
\alias{is_cnpj}
\title{Testa se o valor eh CNPJ}
\usage{
is_cnpj(i, fix_na = TRUE)
}
\arguments{
\item{i}{vetor de CNPJ. Ver \code{\link[validaRA]{valida_doc}}}

\item{fix_na}{vetor de strings a serem removidas caso existam.}
}
\description{
Com apoio do pacote "validaRA" testa se o vetor recebido e um CNPJ.
}
