% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_date}
\alias{is_date}
\title{Testa se o valor eh data}
\usage{
is_date(i, check = TRUE, fix_na = TRUE)
}
\arguments{
\item{i}{vetor qualqer}

\item{check}{se TRUE (padao) avalia o conteudo de i e infere se a variavel total e de data}

\item{fix_na}{se tiver algum tipo de na no vetor sera removido}
}
\value{
TRUE se o vetor e das classes de data
}
\description{
Permite avaliar, atraves do tipo ou do conteudo, se um vetor de dados recebido e ou nao no tipo data
}
