% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_email}
\alias{is_email}
\title{Verifica se a string eh email}
\usage{
is_email(email, rm_accent = TRUE, fix_na = TRUE)
}
\arguments{
\item{email}{vetor de characters.}

\item{rm_accent}{logico indicando se deve buscar e remover acentos em email}

\item{fix_na}{se tiver algum tipo de na no vetor sera removido}
}
\value{
Vetor de characters detectados como e-mail
}
\description{
Aplica-se uma regex bem formatada para checar se as strings contendo dados de e-mails
esta correta. Se houver alguma alguma que nao bate com a regex entao esta e removida
e tratada como nao e-mail ou e-mail invalido.
}
