% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_num}
\alias{is_num}
\title{Testa se o valor eh numerico}
\usage{
is_num(i, check = TRUE, fix_na = FALSE)
}
\arguments{
\item{i}{vetor qualqer numerico}

\item{check}{se TRUE (padao) avalia o conteudo de i e infere se a variavel e de data}

\item{fix_na}{se tiver algum tipo de NA no vetor sera removido}
}
\value{
TRUE se o vetor i e das classes numericas
}
\description{
Permite avaliar, atraves do tipo ou do conteudo, se um vetor de dados recebido e ou nao no tipo numerico
}
