% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{numeric_precision_and_scale}
\alias{numeric_precision_and_scale}
\title{Calculates the biggest precision and scale that occurs in a numeric vector}
\usage{
numeric_precision_and_scale(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A list with four elements:
        precision (total number of significant digits in the whole number),
        scale (number of digits in the fractional part),
        non.fractional.precision (number of digits at the left side and SQL precision.
}
\description{
CREDITOS: https://stackoverflow.com/questions/42122215
The scale of a numeric is the count of decimal digits in the fractional part (to the right of the decimal point).
The precision of a numeric is the total count of significant digits in the whole number,
that is, the number of digits to both sides of the decimal point.

To create a suitable numeric data type in a SQL data base use the returned \code{SQL.precision} which
is defined by \code{max(precision, non.fractional.precision + scale)}.
}
\details{
NA will be counted as precision 1 and scale 0!
}
\examples{
x <- c(0, 7654321, 54321.1234, 321.123, 321.123456789, 54321.1234, 100000000000, NA)
numeric_precision_and_scale(x)
numeric_precision_and_scale(c(10.0, 1.2))
}
