% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{rnp_2freq}
\alias{rnp_2freq}
\title{Tabela de frequencias dupla entrada.}
\usage{
rnp_2freq(data, x, y = NULL, sortclass = FALSE, transpor = FALSE)
}
\arguments{
\item{data}{tabela de dados.}

\item{x}{nome da variavel x. Obrigatorio.}

\item{y}{nome da variavel y. Pode ser NULL.}

\item{sortclass}{Se TRUE ordena as categorias de x e/ou y.}

\item{transpor}{Se TRUE, faz a transposicao das estatisticas gerando um data.frame vertical.}
}
\value{
data.frame
}
\description{
Tabela de frequencias dupla entrada.
}
\details{
faz tabela de dupla entrada com base no dados de entrada.
}
\examples{
\dontrun{
require(rnpextras)
data(mtcars)
rnp_2freq(mtcars, "am")
rnp_2freq(mtcars, "am", transpor = TRUE)
rnp_2freq(mtcars, "am", "cyl")
rnp_2freq(mtcars, "am", "cyl", transpor = TRUE)
}
}
\author{
LOPES, J. E
}
