% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/extras.R
\name{rnp_associacao}
\alias{rnp_associacao}
\title{Estatisticas de associacao}
\usage{
rnp_associacao(x, y, ...)

rnp_associacao(x, y, ...)
}
\arguments{
\item{x}{variavel um}

\item{y}{variavel dois}

\item{...}{argumentos passados para a funcao \code{\link{chisq.test}}}
}
\description{
Funcao para calcular as estatisticas, Qui-Quadrado, V de Cramer e C de contingencia para tabelas de dupla entrada.

Recebe dois vetores categoricos de tamanhos iguais e calcula as estatisticas
Qui-quadrado, V de Cramer e o coeficiante de contingencia para a tabela de contingencia
das duas variaveis criada pela funcao \code{\link{table}}.
}
\examples{
require(rnpextras)
rnp_associacao(dm_ies$DESC_TP_ORGANIZACAO_ACADEMICA, dm_ies$DESC_TP_CATEGORIA_ADMINISTRATIVA)
}
\author{
LOPES, J. E.
}
