% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_aux_imputation}
\alias{rnp_aux_imputation}
\title{Funcao para tratar uma coluna contendo NAs}
\usage{
rnp_aux_imputation(x, fun = "median")
}
\arguments{
\item{x}{Coluna que contem NAs e necessita ser corrigida}

\item{fun}{Tipo de tratamento a ser realizado, onde a entrada deve ser um dos seguintes: ("mean","median","min","max","tsclean")}
}
\description{
Esta funcao realiza o tratamento de NAs de um vetor onde o tratamento depende do parametro de entrada "fun"
}
\seealso{
\code{\link{tsclean}}
}
\author{
LOPES, J. E.; PLACIDO, V. B.
}
