% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_check_class_vars}
\alias{rnp_check_class_vars}
\title{Converte e categoriza numerico}
\usage{
rnp_check_class_vars(x, min_categ = 50, nfaixas = 20, coarse_class = TRUE)
}
\arguments{
\item{x}{vetor de valores}

\item{min_categ}{minimo de categorias para que a variavel seja recategorizada}

\item{nfaixas}{numero de faixas desejadas}

\item{coarse_class}{se TRUE tenta converter x para numerico antes de recategorizar.}
}
\description{
Verifica se uma variavel tem muitas classes, converte para numerica e categoriza
}
