% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_check_day}
\alias{rnp_check_day}
\title{Dia da semana}
\usage{
rnp_check_day(x, en = TRUE)
}
\arguments{
\item{x}{vetor de datas}

\item{en}{se quer ver em ingles}
}
\description{
Funcao para determinar o dia da semana e separar os periodos de extracao de dados
}
