% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_clean_brackets}
\alias{rnp_clean_brackets}
\title{Remove colchetes, parentesis e chaves}
\usage{
rnp_clean_brackets(x)
}
\arguments{
\item{x}{string de dados}
}
\description{
Remove colchetes, parentesis e chaves
}
