% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_cut}
\alias{rnp_cut}
\title{Particao de dados}
\usage{
rnp_cut(n, nrows = 1000)
}
\arguments{
\item{n}{Total de linhas da base total}

\item{nrows}{Total de linhas por bloco}
}
\description{
Ajuda a quebrar um conjunto de dados com base no total de linhas desejado.
}
