% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_db_read}
\alias{rnp_db_read}
\title{Leitura de dados do banco via ODBC}
\usage{
rnp_db_read(con, query, chunk_size = 24999, progress = TRUE)
}
\arguments{
\item{con}{Conexao ativa odbc. Veja \code{\link[DBI]{dbConnect}} para mais detalhes.}

\item{query}{Consulta SQl que retorna dados.}

\item{chunk_size}{Total de linhas extraidas por batch.}

\item{progress}{Acompanha log da extracao.}
}
\description{
Recebe uma conexao odbc criada por \code{\link[DBI]{dbConnect}} e passa uma query
para ober os resultados. Aceita todo tipo de consulta compativel com SQL server, inclusive
chamadas de execucao de procedures que retornam dados.
}
