% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_db_write}
\alias{rnp_db_write}
\title{Grava dados em um banco via ODBC}
\usage{
rnp_db_write(
  con,
  data,
  name,
  column_types = NULL,
  schema = "dbo",
  method = "dbi",
  chunk_size = 24999,
  verbose = TRUE,
  append = FALSE
)
}
\arguments{
\item{con}{Conexao ativa odbc. Veja \code{\link[DBI]{dbConnect}} para mais detalhes}

\item{data}{Objeto de dados. Pdoe ser data.frame ou qualquer objeto que se posssa converter em data.frame}

\item{name}{String de nome da tabela que vai ser grabada no banco com os dados do objeto 'data'}

\item{column_types}{Tipagem das colunas. Se 'auto' salva com padroes do pacote DBI, se 'otima' busca
padros conforme estrutura dos dados. Pode ser muito demorado sem bases grandes. Se 'manual',
o usuario deve informar um vetor nomeado onde cada nome e uma coluna da base e o valor e o
formato compativel com o SQL Server.}

\item{schema}{String de nome do schema do banco de dados que deseja salvar. padrao e 'dbo'}

\item{method}{String com um entre os valoes 'dbi' ou 'copy'. Vaje details para mais informacoes.}

\item{chunk_size}{Define o total de linhas que cada lote salvara no banco por vez.}

\item{verbose}{Se TRUE que mostra todo o log do processo ou apenas algumas partes.}

\item{append}{Se TRUE e a tabela ja existir no banco de dados ele tenta apendar (empilhar) os dados.}
}
\description{
Recebe parametros de uma conexao odbc ativa e exporta para o respectivo banco de dados
uma tabela de dados do R como data.frame, tibbla, matriz ou data.table subindo inteira ou por partes.
}
\details{
Quando o argumento column_types e NULL, padrao, a configuracao dos tipos de dados e feita
automaticamente atraves de metodos do pacote DBI, dplyr e dbplyr. Porem, o usuario pode passar
manualmente um vetor nomeado com os tipos compativeis com SQL. Cada valor e um tipo e cada nome
e uma coluna da tabela. Para obter as tipagens de forma otimizada, criamos a funcao \code{\link{rnp_db_types}}.
Esta funcao pode ser muito lenta, pois ela analisa a fundo a estrutura dos dados de toda as colunas
para determinar o tipo exato, mas permite paralelismo. Optamos por manter esta funcao fora das rotinas
da \code{\link{rnp_db_write}} devido seu carater de computacao custoso, mas recomendamos trabalhar com
ela para detectar corretamente so tipos exatos antes de criar a tabela fisica no banco de dados.
Existem duas formas de salvar dados no banco de dados: se a escolha for
method = 'dbi' o R tentara salvar a base de dados em partes utilizando methodos do pacote DBI,
caso method = 'copy' ele tentara salvar os dados de uma unica vez sem quebra em partes.
Este metodo pode evitar problemas com tipos de dados, mas costuma ser custoso computacionalmente.
Para mais detalhes ver \code{\link[DBI]{dbWriteTable}}, \code{\link[dplyr]{copy_to}} e
\code{\link[dbplyr]{in_schema}}.
Caso o usuario passe uma tipagem manual, ele precisa passar para todas as variaveis da tabela.
Exemplo, se a tabela tiver 3 colunas, column_types = c(mpg = 'DECIMAL(3,1)', cyl = 'VARCHAR(20)', logic = 'BIT') .
}
