% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_descritiva}
\alias{rnp_descritiva}
\title{Analise descritiva}
\usage{
rnp_descritiva(x, digits = 4)
}
\arguments{
\item{x}{vetor numerico}

\item{digits}{cadas decimais no arredondamento}
}
\description{
Calcula estatisticas descritivas para variaveis numericas.
}
