% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_freq}
\alias{rnp_freq}
\title{Tabela de frequencias simples}
\usage{
rnp_freq(x, digits = 4, percent = FALSE, sortclass = TRUE)
}
\arguments{
\item{x}{varivetor de entrada}

\item{digits}{arredondar as frequencias para...}

\item{percent}{se TRUE retorna os percentuais formatados como \%}

\item{sortclass}{ordenacao das classes.}
}
\value{
data.frame com as estaisticas frequencia absoluta (fa), relativa (fr)
absoluta acumulada (faa) e relativa acumulada (fra)
}
\description{
Calcula estatisticas descritivas de variaveis categoricas como caractere
e fatores. Tambem funciona com numericas com valores repetidos. Como resultado tem-se
frequencias e percentuais simples e acumulados de cada categoria.
Se receber um data.frame tenta calcular as estatisticas para todas as variaveis da mesma.
Nao existe deduplicacao dos dados.
}
\examples{
rnp_freq(mtcars$cyl)
rnp_freq(mtcars$cyl, percent = TRUE)
rnp_freq(mtcars, percent = TRUE)
}
\author{
LOPES, J. E.
}
