% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_gains_table_score}
\alias{rnp_gains_table_score}
\title{Gains table para o score}
\usage{
rnp_gains_table_score(
  dados,
  variavel_mau = "status_id_atual",
  variavel_score = "score",
  variavel_valor = NULL,
  nquebras = 20,
  define_mau = c(0, "M", "MAU", "BAD"),
  digits = 4,
  credito = FALSE,
  completa = TRUE,
  exclusoes = NULL
)
}
\arguments{
\item{dados}{base de dados com dados com campos necessarios para construir a tabela}

\item{variavel_mau}{nome da variavel que contem os codigo de status dos maus.}

\item{variavel_score}{nome do campo que possui o score. Deve ser numerico.}

\item{variavel_valor}{nome do campo com valor total em unidades monetárias, se houver. Deve ser numerico. Se nao tiver deixe NULL, padrao.}

\item{nquebras}{total de quebras da base de dados com base na quantidade de pedidos (volume)}

\item{define_mau}{vetor de codigos de status a considerar como mau. Padrao c(0, "M", "MAU", "BAD")}

\item{digits}{total de digitos para as estatisticas na tabela de saida}

\item{credito}{Se TRUE, gera tabela em ordem inversa, ou seja maiores escores sao melhores que menores scores. O processo inverso da tabela se aplica a modelos de scores de fraude.}

\item{completa}{TRUE para determinar o tabelao com informacoes de valores monetarios}

\item{exclusoes}{vetor numerico de codigos de status a excluir da base de dados}
}
\description{
Gera uma tabela de ganho do score com muitas estatísticas úteis para a análise
dos dados de modelos de score baseados em regressão com dados binários.
}
