% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_gains_table_variavel}
\alias{rnp_gains_table_variavel}
\title{Gains table visão variavel}
\usage{
rnp_gains_table_variavel(
  dados,
  variavel_mau = "status_id_atual",
  variavel_categ = "reg_combin",
  digits = 4,
  exclusoes = NULL,
  define_mau = NULL
)
}
\arguments{
\item{dados}{base de dados com dados com campos necessarios para construir a tabela}

\item{variavel_mau}{nome da variavel que contem os codigo de status.}

\item{variavel_categ}{nome da variavel categórica de interesse.}

\item{digits}{total de digitos para as estatisticas na tabela de saida}

\item{exclusoes}{vetor numerico de codigos de status a excluir da base de dados}

\item{define_mau}{vetor de codigos de status a considerar como mau. Padrao e c(0, "M", "MAU", "BAD")}
}
\description{
Tabela de ganhos para categorias de variaveis presentes ou nao em modelos.
}
