% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_inferencia_rep_parcelamento}
\alias{rnp_inferencia_rep_parcelamento}
\title{Inferencia de performance por Parcelamento}
\usage{
rnp_inferencia_rep_parcelamento(
  dados_apa,
  dados_rep,
  variavel_mau = "status_id_atual",
  variavel_score = "score",
  variavel_valor = NULL,
  nquebras = 20,
  define_mau = c(10, 22, 32, 33, 34, 36, 43, 103, 123, 160, 161, 191, 203, 204),
  digits = 4,
  credito = FALSE,
  exclusoes = NULL,
  return_all = FALSE,
  seed = 1983
)
}
\arguments{
\item{dados_apa}{base de dados dos aprovados, precisa ter score e marcacao de maus para construir o tabelao.}

\item{dados_rep}{base de dados dos aprovados, precisa ter score e marcacao de maus para construir o tabelao.}

\item{variavel_mau}{nome da variavel que contem os codigo de status dos pedidos.}

\item{variavel_score}{nome do campo que possui o score do pedido. Deve ser numerico.}

\item{variavel_valor}{nome do campo com valor total do pedido. Deve ser numerico.
Se nao tiver deixe NULL, padrao.}

\item{nquebras}{total de quebras da base de dados com base na quantidade de pedidos (volume)}

\item{define_mau}{vetor de codigos de status a considerar como mau.
Padrao de modelagem e (10,22,32,33,34,36,43,103,123,160,161,191,203,204)}

\item{digits}{total de digitos para as estatisticas na tabela de saida}

\item{credito}{Se TRUE, gera o tabelao em ordem inversa, ou seja maiores escores
sao melhores que menores scores. O processo inverso do tabelao de scores de fraude.}

\item{exclusoes}{vetor numerico de codigos de status a excluir da base de dados}

\item{return_all}{se FALSE retorna tabelao dos aprovados, tabelao dos inferidos e
base dos reprovados com marcacao dos inferidos. Caso contrario, retorna apenas
tabelao dos inferidos.}

\item{seed}{Padrao 1983. Altamente recomendado alterar essa semente a gosto para
tornar sua amostragem replicavel no futuro.}
}
\value{
Se return_all == TRUE retorna tabelao dos aprovados, tabelao dos inferidos e
base dos reprovados com marcacao dos inferidos. Caso contrario, retorna apenas
tabelao dos inferidos.
}
\description{
Utiliza a tecnica do parcelamento para inferir a performance de uma base de
Cancelados/Rejeitadoscom base na tabela dos Aprovados contendo a marcacao de maus.
A tecnica do parcelamento pode ser compreendida no trabalho de Ash, D e Meester, S. (2002)
 - Best Practices in Reject Inferencing. A estrategia e a seguinte:
\itemize{
\item{gera faixas de percentis dos dados ordenando por score na base de aprovados (tabelao);}
\item{calcula taxa de maus em cada faixa de score;}
\item{marca a base dos Reprovados nas mesmas faixas de score;}
\item{faz amostragem aleatoria simples em cada faixa da base dos Reprovados tomando como
 tamanho da amostra o percentual de maus das faixas obtidas da base dos Aprovados;}
\item{em cada amostra, fixa-se como mau todas as observacoes obidas;}
\item{recompoe a base dos Cancdelados com a identificacao das transacoes que foram marcadas
 como mau na amostragem.}
}
}
\examples{
\dontrun{
require(rnpextras)
out <- rnp_inferencia_rep_parcelamento(dados_apa = dados_apa,
                                       dados_rep = dados_rep,
                                       variavel_mau = "cbk_fraude",
                                       variavel_score = "score",
                                       nquebras = 20,
                                       define_mau = 1,
                                       return_all = TRUE,
                                       seed = 1983)
}
}
