% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_ksp}
\alias{rnp_ksp}
\title{KSP - KS de Proximidade}
\usage{
rnp_ksp(
  tbref,
  tbrec,
  max_cat = 50,
  digits = 4,
  total = FALSE,
  sortclass = TRUE
)
}
\arguments{
\item{tbref}{tabela de frequencias da base de referencia (esperado) calculada com a funcao rnp_freq()}

\item{tbrec}{tabela de frequencias da base de recente (atual) calculada com a funcao rnp_freq()}

\item{max_cat}{maximo de categorias devolvidas}

\item{digits}{digitos para arredondar}

\item{total}{se TRUE retorna o KSP somado, se nao, exibe por cateoria}

\item{sortclass}{se TRUE, ordena as categorias}
}
\value{
se total == TRUE retorn data.frame com percentuais de cada classe da variavel mais estatistica VDI e inferencia do efeito.
Caso contrario, apenas um data.frame com estatistica VDI e inferencia do efeito.
}
\description{
Serve para calcular o indice de desvio entre duas frequencias, sendo a referencia ou esperada versus a recente ou atual
Esta estatistica e calculada com base no KS como conhecemos porem, adaptado para o caso em que
queremos analisar a distancia entre duas frequencias acumuladas e ajuda a identificar se esta havendo mudanca de publico olhando
para os mesmos dados antes e depois ao longo do tempo. A interpretacao e oposta ao KS de modelagem, ou seja, quanto menor-melhor.

Calculamos o KSP como $max(abs(freq_relativa_ac_p1 - freq_relativa_ac_p2))$
}
\examples{
tbref <- rnp_freq(sample(mtcars$cyl, size = 20), digits = 6)
tbrec <- rnp_freq(sample(mtcars$cyl, size = 20), digits = 6)
rnp_ksp(tbref, tbrec, total = TRUE)
rnp_ksp(tbref, tbrec, total = FALSE)
}
\author{
LOPES, J. E.
}
