% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_cep2}
\alias{rnp_limpa_cep2}
\title{Limpa CEP}
\usage{
rnp_limpa_cep2(cep, fix_na = TRUE)
}
\arguments{
\item{cep}{vetor de numeros ou strings de CEP}

\item{fix_na}{padroes de NA removidos do numero ou string i}
}
\description{
Via regex e limpeza de dados, verifica-se a estrutura do vetor informado, removendo tudo que nao for
numero e completando com zeros a esquerda casos com menos de 8 digitos. Se o CEP for da forma
99999-999, o traco e removido e o numero retorna limpo.
nao validamos se o cep e valido nos correios.
}
