% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_cpf_cnpj}
\alias{rnp_limpa_cpf_cnpj}
\title{Limpa CPF/CNPJ}
\usage{
rnp_limpa_cpf_cnpj(doc, fix_na = TRUE)
}
\arguments{
\item{doc}{vetor de numeros ou strings de CPF/CNPJ.}

\item{fix_na}{padroes de NA removidos do numero ou string i}
}
\value{
tibbla com tres colunas, uma com o doc outras com logico se CPF = TRUE ou CNPJ = TRUE
}
\description{
Com apoio do pacote "validaRA", regexes e avaliacao do comprimento das strings no campo de
CPF/CNPJ recebido, verifica a validade do campo segundo regras de construcao do CPF/CNPJ e acerta o tamanho.
Pontos, barras e tracos sao removidos e apenas um vetor numerico em forma de string e retornado.
Ver \code{\link[validaRA]{valida_doc}} para mais detalhes.
}
