% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_date}
\alias{rnp_limpa_date}
\title{Limpa Data}
\usage{
rnp_limpa_date(date, fix_na = TRUE, format_out = NULL, char_out = FALSE)
}
\arguments{
\item{date}{Vetor de data generico.}

\item{fix_na}{Padroes de NA removidos do numero ou string i}

\item{format_out}{Se desejar, pode informar um formato de data para a saida. Veja \code{\link[base]{format.Date}}}

\item{char_out}{Se TRUE, retorna a data formatada como char.}
}
\description{
Valida 27 padroes de combinacoes de datas incluindo data simples e datetime.
}
\details{
Esta funcao trabalha com timezone = UTC: hora universal. A localizacao e definida
pelo padrao do sistema. Ver \code{\link[lubridate]{parse_date_time}} e \code{\link[base]{locales}}.
}
\examples{
\dontrun{
x <- c('February 20th 1973',
'Sunday, May 1, 2000',
'february  14, 04',
'Feb 20th 73',
'January 5 1999 at 7pm',
'jan 3 2010',
'Jan 1, 1999',
'jan 3   10',
'01 3 2010',
'1 3 10',
'1 13 89',
'5/27/1979',
'12/31/99',
'DOB:12/11/00',
'Thu, 1 July 2004 22:30:00',
'Thu, 1st of July 2004 at 22:30:00',
'Thu, 1July 2004 at 22:30:00',
'Thu, 1July2004 22:30:00',
'Thu, 1July04 22:30:00',
'21 Aug 2011, 11:15:34 pm',
'1979-05-27 05:00:59',
'1979-05-27',
'3 jan 2000',
'17 april 85',
'27/5/1979',
'20 01 89',
'00/13/10',
'14 12 00',
'03:23:22 pm')
rnp_limpa_date(x)
}
}
