% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_email}
\alias{rnp_limpa_email}
\title{Limpa texto de e-mail baseado em regex}
\usage{
rnp_limpa_email(email, rm_accent = TRUE, fix_na = TRUE)
}
\arguments{
\item{email}{vetor de characters.}

\item{rm_accent}{logico indicando se deve buscar e remover acentos em email}

\item{fix_na}{se tiver algum tipo de na no vetor "email" sera removido}
}
\value{
Vetor de characters detectados como e-mail
}
\description{
Aplica-se uma regex bem formatada para checar se as strings contendo dados de e-mails
estao corretas. Se houver alguma alguma string que nao bate com a regex entao esta e removida
e tratada como nao e-mail ou e-mail invalido.
}
