% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_tel}
\alias{rnp_limpa_tel}
\title{Corrige numero de telefone}
\usage{
rnp_limpa_tel(tel)
}
\arguments{
\item{tel}{vetor de numeros de celular}
}
\description{
Recebe um vetor com numeros de telefone e tentar ajustar
para o padrao limpo aceitavel. Forma adaptada para rodar tambem em data.frame
e matriz
}
\details{
o objeto tel tera qualquer digito diferente de zero removido.
verifica-se o tamenho do numero e adicona DDI;
}
