% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_tel_boost}
\alias{rnp_limpa_tel_boost}
\title{Limpa tel boost}
\usage{
rnp_limpa_tel_boost(tel)
}
\arguments{
\item{tel}{vetor de numeros de tel}
}
\description{
Limpa tel boost
}
\details{
o objeto cep tera qualquer digito diferente de zero removido.
}
