% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_limpa_text}
\alias{rnp_limpa_text}
\title{Limpa texto textos em geral}
\usage{
rnp_limpa_text(
  text,
  char_size = "n",
  regex_from = NULL,
  regex_to = NULL,
  rm_accent = TRUE,
  fix_na = TRUE
)
}
\arguments{
\item{text}{vetor de characters.}

\item{char_size}{'n' para normal, nada (nothing), 'l' para minuscula (lower) e 'u' para maiuscula (upper).}

\item{regex_from}{Aceita regex para buscar e substituir em text.}

\item{regex_to}{Valor para substituir em regex_from. Se NULL, substitui por vazio "".}

\item{rm_accent}{logico indicando se deve buscar e remover acentos em email}

\item{fix_na}{se tiver algum tipo de na no vetor "email" sera removido}
}
\value{
Vetor de characters detectados como e-mail
}
\description{
Util para limpar espacos em branco no meio das palavras, quebras de linha,
pagina entre outros, se definido, remove acentos e devolve em maiuscula ou minucula.
}
