% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_odds_shift}
\alias{rnp_odds_shift}
\title{Calcula os testes de Odds Shift e Qui-Quadrado}
\usage{
rnp_odds_shift(tab, i = TRUE)
}
\arguments{
\item{tab}{tabela contendo distribuicoes nomeadas baseline (x) e recente (y)}

\item{i}{parametro para bootstrap (uso interno)}
}
\description{
Recebe uma tabela contendo as distribuicoes de frequencias nomeadas
baseline (x) e recent (y) e calcula o teste chi-quadrado para homogeneidade
das classes de cada variavel sob as hipoteses:
H_0 = a distribuicao das frequencias sao iguais (nao diferem)
Se o p-value > 0.95 aceita H_0
}
\details{
Note que existe um argumento i, isso a torna compativel com simulacoes por bootstrap
gerando intervalos de confianca para o indicador.
}
\examples{
\dontrun{
require(boot)
base <- mtcars[,c("gear","vs")]
colnames(base) <- c("var_fx", "y")
a <- rnp_tabulate(dados = base, vx = "var_fx", vy = "y", bom = 0, perf = TRUE)
a <- a[,c("p_bons", "p_maus")]
colnames(a) <- c("x","y")
rnp_odds_shift(a)
boot::boot(data = a, statistic = rnp_odds_shift, R = 50)
}
}
