% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_psi}
\alias{rnp_psi}
\title{Calcula Population Stability Index / Variable Deviation Index (PSI/VDI)}
\usage{
rnp_psi(tab, i = TRUE, cat = FALSE)

rnp_psi(tab, i = TRUE, cat = FALSE)
}
\arguments{
\item{tab}{tabela contendo distribuicoes nomeadas baseline (x) e recente (y)}

\item{i}{parametro para bootstrap (uso interno)}

\item{cat}{O retorno pode ser total ou por categoria}
}
\description{
Recebe uma tabela contendo as distribuicoes de frequencias de aseline (x)
e recent (y) e calcula o PSI/VDI

Recebe uma tabela contendo as distribuicoes de frequencias de aseline (x) e calcula o KS.
}
\details{
Note que existe um argumento i, isso a torna compativel com simulacoes por bootstrap
gerando intervalos de confianca para o indicador.

Note que existe um argumento i, isso a torna compativel com simulacoes por bootstrap
gerando intervalos de confianca para o indicador.
}
\author{
LOPES, J. E.
}
