% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_read_bf_dt}
\alias{rnp_read_bf_dt}
\title{Leitura de dados com pacote data.table}
\usage{
rnp_read_bf_dt(file, control = rnp_read_control(type = "fread"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessarios para importar dados para uso inteno com a rnp_read()}
}
\description{
Este e apenas um wraper que deixa a funcao \code{\link[data.table]{fread}}
melhor formatada para passar na funcao \code{\link{rnp_read}}
}
