% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_read_bf_rd}
\alias{rnp_read_bf_rd}
\title{Leitura de dados com pacote readr}
\usage{
rnp_read_bf_rd(file, control = rnp_read_control(type = "readr"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessarios para importar dados para uso inteno com a rnp_read()}
}
\description{
Este e apenas um wraper que deixa as funcoes \code{\link[readr]{read_csv2}} e
\code{\link[readr]{read_delim}} conveniente para passar na funcao \code{\link{rnp_read}}
}
