% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_read_bf_rio}
\alias{rnp_read_bf_rio}
\title{Leitura de dados com pacote rio}
\usage{
rnp_read_bf_rio(file, ...)
}
\arguments{
\item{file}{Nome do arquivo a ser importado}

\item{...}{Argumentos extras}
}
\description{
Este e apenas um wraper que deixa a funcao \code{\link[rio]{import}}
melhor formatada para passar na funcao \code{\link{rnp_read}}
}
