% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_read_control}
\alias{rnp_read_control}
\title{Controle para leitura de arquivos grandes (Big Files)}
\usage{
rnp_read_control(
  type = "fread",
  delim = NULL,
  dec = NULL,
  try_append = TRUE,
  sep = NULL,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_select = NULL,
  col_types = NULL,
  encoding = NULL,
  na = c("", "NA"),
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  altrep = TRUE,
  progress = NULL,
  skip_empty_rows = TRUE,
  ...
)
}
\arguments{
\item{type}{Nome do mecanisco de leitura. Podendo sem um em "fread","readr","vroom" ou "rio". Veja os detalhes abaixo para saber mais.}

\item{delim}{Delimitador de colunas, se nao souber deixe NULL. Pode ser, ",", ";", "|" e outros conforme seu arquivo.}

\item{dec}{Separador decimal.}

\item{try_append}{Se TRUE (padrao) tenta apendar (empilhar) os dados. Se FALSE retorna lista de tabelas, uma para cada arquivo importado.}

\item{sep}{Agrupador de campos para formatacao de numeros}

\item{quote}{Regex para quotation. Serve para nao confundir com aspas que delimita strings.}

\item{escape_backslash}{Utilizar contrabarra para saltar Caracteres especiais}

\item{escape_double}{Se TRUE o valor '""' representara um unico \"}

\item{col_names}{TRUE, FALSE, um vetor de nomes ou numeros que serao tratados como indices para selecionar colunas}

\item{col_select}{Vetor de nomes ou indices a selecionar.}

\item{col_types}{Se NULL trabalha com os padroes do pacote readr. Se data.table, utiliza os padroes da funao \code{\link[data.table]{fread}}}

\item{encoding}{Encoding do arquivo. Como padrao pega as configuracoes do Sistema Operacional}

\item{na}{Tipo de valor para substituir NA}

\item{comment}{Caractere a ser tratado como comentario. padrao e hashtag}

\item{trim_ws}{Limpar espacos em branco a direita e esquerda das colunas}

\item{skip}{numero de linhas a saltar na leitura dos dados}

\item{n_max}{numero de linhas que quer extrair. Se Inf retorna o maximo de linhas}

\item{guess_max}{Total de linhas por partes lidas (chunksize)}

\item{altrep}{Controle dos tipos de dados na leitura por \code{\link[vroom]{vroom}}}

\item{progress}{Mostrar progresso da leitura dos dados}

\item{skip_empty_rows}{Ignorar linhas vazias}

\item{...}{Argumentos extras}
}
\description{
Esta funcao e uma chave para passar argumentos para a funcao rnp_read
}
