% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_resumo}
\alias{rnp_resumo}
\title{Resumo de dados}
\usage{
rnp_resumo(x)
}
\arguments{
\item{x}{vetor, data.frame ou lista de data.frames}
}
\description{
Recebe um objeto qualquer e retorna um resumo dos dados dentro deste objeto
}
\details{
O resumo contém dados descritivos, classes, tipos, numero de missing, etc.
}
