% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_rm_accent}
\alias{rnp_rm_accent}
\title{Remove acentos de caracteres.}
\usage{
rnp_rm_accent(str)
}
\arguments{
\item{str}{vetor de characters.}
}
\value{
Vetor de characters sem acentos e sem cedilhas.
}
\description{
Remove acentos agudos, circunflexos, tremas, crases e cedilhas de Caracteres e strings.
}
\examples{
\dontrun{
palavras <- c("\u00e1gua", "op\u00e7\u00e3o", "\u00ff\u00c0\u00ea")
rnp_rm_accent(palavras)
}
}
