% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_rm_na}
\alias{rnp_rm_na}
\title{Substitui NA numerico ou char.}
\usage{
rnp_rm_na(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{Vetor qualquer}

\item{from}{Qual o codigo que quer remover.}

\item{to}{Qual o valor a substituir}
}
\value{
Vetor x com as devidas alteracoes
}
\description{
Substitui NA por algum valor fornecido ou padroniza com NA.
}
\details{
Quando from = NULL, padrao, busca em numerico os valores  'NA, NaN, NULL, -Inf, Inf' e
quando character busca por '','.','NA','na','N/A','n/a','NaN','nan' e troca pelo valor de
to. Quando to = NULL, padrao, substitui pelo padrao do R que e NA.
}
