% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_shift2}
\alias{rnp_shift2}
\title{Shift}
\usage{
rnp_shift2(x, y, digits = 4)
}
\arguments{
\item{x}{vetor x}

\item{y}{vetor y}

\item{digits}{total de digitos do arredondamento}
}
\description{
Calcula o shift ou mudanca e pode ser aplicado a qualquer par de vetores.
}
