% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_shift_cv}
\alias{rnp_shift_cv}
\title{Calcula shift do coeficiente de variacao}
\usage{
rnp_shift_cv(p1, p2, digits = 4, gatilho = NULL)
}
\arguments{
\item{p1}{Vetor de dados do periodo 1}

\item{p2}{Vetor de dados do periodo 2}

\item{digits}{Arredondamento do numero}

\item{gatilho}{Se preferiri, passar manualmente qual o gatilho para dizer se um shift e alerta. ex. 0.22}
}
\description{
Dados dois vetores numericos com dados de dois periodos, p1 e p2 fazemos o coeficiente de variacao (cv)
de cada vetor e cria uma estatistica de shift entre os dois cv's.
O cv de cada daso e dado por cvp1 = sd(p1) / mean(p1) e cvp1 = sd(p2) / mean(p2). O shift e dado por
shift = (cvp2-cvp1) / cvp1. Se shift <= 0.01, alerta = 0, se > 0.05 e <= 0.10, alerta = 1,
se > 0.10 e <= 0.15, alerta = 2 e segue ate o 4.
Testar se alertas a partir de 1 fazem sentido. Esta regua pode ser calibrada.
}
\author{
LOPES, J. L.
}
