% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_split_data_table}
\alias{rnp_split_data_table}
\title{Split tabela de dados por linhas}
\usage{
rnp_split_data_table(x, nrows = 5000, keep_group = FALSE, append = FALSE)
}
\arguments{
\item{x}{tabela de dados 2 x 2}

\item{nrows}{numero de linhas desejado por subtabela}

\item{keep_group}{deseja que a coluna que identifica os lotes seja retornada?
padrao FALSE.}

\item{append}{se TRUE, mantem campo keep_group e junta os dados gerando uma unica tabela. Padrao = FALSE}
}
\value{
lista de subtabelas
}
\description{
Recebe uma tabela das classes "data.frame","data.table","tbl_df", "tbl"
e quebra em uma lista de subtabela de acordo com o total de linhas desejado. Se na
ultima parte nao houber linhas suficientes para fecar nrow, retorna o resto das linhas.
}
