% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_stats}
\alias{rnp_stats}
\title{Estatisticas descritivas mais completas}
\usage{
rnp_stats(x, digits = 4, conf.level = 0.95)
}
\arguments{
\item{x}{vetor numerico}

\item{digits}{cadas decimais no arredondamento}

\item{conf.level}{confianca para o intervalo}
}
\description{
Calcula estatisticas descritivas e intervalo de confianca para a media
}
