% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_tabelao_score}
\alias{rnp_tabelao_score}
\title{Tabelao score}
\usage{
rnp_tabelao_score(
  dados,
  variavel_mau = "status_id_atual",
  variavel_score = "score",
  variavel_valor = NULL,
  nquebras = 20,
  define_mau = c(10, 22, 32, 33, 34, 36, 43, 103, 123, 160, 161, 191, 203, 204),
  digits = 4,
  credito = FALSE,
  completa = TRUE,
  exclusoes = NULL
)
}
\arguments{
\item{dados}{base de dados com dados com campos necessarios para construir o tabelcao}

\item{variavel_mau}{nome da variavel que contem os codigo de status dos pedidos. E.: pedido_status_id. Deve ser numerico.}

\item{variavel_score}{nome do campo que possui o score do pedido. Deve ser numerico.}

\item{variavel_valor}{nome do campo com valor total do pedido. Deve ser numerico. Se nao tiver deixe NULL, padrao.}

\item{nquebras}{total de quebras da base de dados com base na quantidade de pedidos (volume)}

\item{define_mau}{vetor de codigos de status a considerar como mau. Padrao de modelagem e (10,22,32,33,34,36,43,103,123,160,161,191,203,204)}

\item{digits}{total de digitos para as estatisticas na tabela de saida}

\item{credito}{Se TRUE, gera o tabelao em ordem inversa, ou seja maiores escores sao melhores que menores scores. O processo inverso do tabelao de scores de fraude.}

\item{completa}{TRUE para determinar o tabelao com informacoes de FRD (22), CBK(32, 33, 34, 36, 43, 103) e SUS (10)}

\item{exclusoes}{vetor numerico de codigos de status a excluir da base de dados}
}
\description{
Tabelao para faixas de score visao volume de pedidos.
}
