% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_tabelao_variavel}
\alias{rnp_tabelao_variavel}
\title{Tabelao variavel}
\usage{
rnp_tabelao_variavel(
  dados,
  variavel_mau = "status_id_atual",
  variavel_categ = "reg_combin",
  digits = 4,
  exclusoes = NULL,
  define_mau = NULL
)
}
\arguments{
\item{dados}{base de dados com dados com campos necessarios para construir o tabelcao}

\item{variavel_mau}{nome da variavel que contem os codigo de status dos pedidos. E.: pedido_status_id. Deve ser numerico.}

\item{variavel_categ}{nome da variavel de interesse}

\item{digits}{total de digitos para as estatisticas na tabela de saida}

\item{exclusoes}{vetor numerico de codigos de status a excluir da base de dados}

\item{define_mau}{vetor de codigos de status a considerar como mau. Padrao e c(0, "M", "MAU", "BAD")}
}
\description{
Tabelao para categorias de variaveis presentes ou nao em modelos.
}
