% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_tamamostra_prop}
\alias{rnp_tamamostra_prop}
\title{Tamanho da amostra por proporcao}
\usage{
rnp_tamamostra_prop(N, p = NULL, z = NULL, e = NULL, infinito = FALSE)
}
\arguments{
\item{N}{Valor do tamanho da base de dados}

\item{p}{E a proporcao que esperamos encontrar. ex 50\% = 1/2 (padrao)}

\item{z}{E o desvio do valor medio que aceitamos para alcancar o nivel de confianca desejado. (ex. 25\%)}

\item{e}{E a margem de erro maximo que eu quero admitir (p.e. 5\%)}

\item{infinito}{se TRUE assume que a populacao e infinita e nao aplica fator de correcao}
}
\value{
vetor com tamanho da amostra n
}
\description{
Calcula tamanho da amostra finita ou infinita baseado em proporcao
N = Tamanho do universo finita
Z = E o desvio do valor medio que aceitamos para alcancar o nivel de confianca desejado.
Em funcao do nivel de confianca que buscamos, usaremos um valor determinado que e dado
pela forma da distribuicao de Normal. Os valores mais frequentes sao:
Nivel de confianca 90\% -> Z=1,645
Nivel de confianca 95\% -> Z=1,96
Nivel de confianca 99\% -> Z=2,575
e = E a margem de erro maximo que eu quero admitir (p.e. 5\%)
p = E a proporcao que esperamos encontrar. Este parametro tende confundir
bastante a primeira vista: Como vou saber qual proporcao espero,
se justamente estamos fazendo uma pesquisa para conhecer esta proporcao?
}
\author{
LOPES, J. L
}
