% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_vdi}
\alias{rnp_vdi}
\title{VDI - Variable Deviation Index}
\usage{
rnp_vdi(
  tbref,
  tbrec,
  max_cat = 50,
  digits = 4,
  total = FALSE,
  sortclass = TRUE
)
}
\arguments{
\item{tbref}{tabela de frequencias da base de referencia (esperado) calculada com a funcao rnp_freq()}

\item{tbrec}{tabela de frequencias da base de recente (atual) calculada com a funcao rnp_freq()}

\item{max_cat}{maximo de categorias devolvidas}

\item{digits}{digitos para arredondar}

\item{total}{se TRUE retorna o VDI somado, senao exibe por cateoria}

\item{sortclass}{se TRUE, ordena as categorias}
}
\value{
se total == TRUE retorn data.frame com percentuais de cada classe da variavel mais estatistica VDI e inferencia do efeito.
Caso contrario, apenas um data.frame com estatistica VDI e inferencia do efeito.
}
\description{
Serve para calcular o indice de desvio entre duas frequencias, sendo a referencia ou esperada versus a recente ou atual
Esta estatistica possui otimas propriedades que ajudam a identificar se esta havendo mudanca de publico olhando
para os mesmos dados antes e depois ao longo do tempo.
}
\examples{
tbref <- rnp_freq(sample(mtcars$cyl, size = 20), digits = 6)
tbrec <- rnp_freq(sample(mtcars$cyl, size = 20), digits = 6)
rnp_vdi(tbref, tbrec, total = TRUE)
rnp_vdi(tbref, tbrec, total = FALSE)
}
\author{
LOPES, J. E.
}
