% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_verifica_celular}
\alias{rnp_verifica_celular}
\title{Limpa celular}
\usage{
rnp_verifica_celular(base, telefone)
}
\arguments{
\item{base}{base de dados}

\item{telefone}{vetor de numeros de tel}
}
\description{
Limpa celular
}
\details{
o objeto cep tera qualquer digito diferente de zero removido.
}
