% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_verifica_cep}
\alias{rnp_verifica_cep}
\title{Corrige CEP}
\usage{
rnp_verifica_cep(base, cep)
}
\arguments{
\item{base}{base de dados}

\item{cep}{vetor de numeros de CEP}
}
\description{
Recebe um vetor com numeros de CEP e tentar ajustar
para o padrao limpo aceitavel.
}
\details{
o objeto cep tera qualquer digito diferente de zero removido.
verifica-se o tamanho do numero e se for menor que oito adiciona zeros a esquerda;
}
