% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_verifica_cnpj}
\alias{rnp_verifica_cnpj}
\title{Limpa CNPJ}
\usage{
rnp_verifica_cnpj(base, cnpj)
}
\arguments{
\item{base}{base com coluna de cnpj}

\item{cnpj}{objetos com os numeros de cpf}
}
\description{
Recebe um vetor de numeros ou textos de CPF's remove qualquer
caractere que nao seja numero e completa com zeros as esquerda. Completa o numero
formando um numero com 14 digitos. Versao vetorizada.
}
