% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rnp_verifica_cpf}
\alias{rnp_verifica_cpf}
\title{Limpa CPF}
\usage{
rnp_verifica_cpf(base, cpf)
}
\arguments{
\item{base}{base com coluna de cpf}

\item{cpf}{objetos com os numeros de cpf}
}
\description{
Recebe um vetor de numeros ou textos de CPF's remove qualquer
carcatere que nao seja numero e completa com zeros as esquerda se tiver
menos de 11 digitos. Caso o valor informado tenha mais de 11 digitos
trata como cnpf e completa formando um numero com 14 digitos. Versao vetorizada.
Se for passado um data.frame, trata cada coluna como uma variavel de CPF e
trata.
}
