% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{try_error}
\alias{try_error}
\title{Trata erro em codigos executados}
\usage{
try_error(code, silent = TRUE)
}
\arguments{
\item{code}{codigo, valor ou operacao}

\item{silent}{se TRUE roda silencionamente}
}
\description{
forma mais inteligente da try(). Com ela o erro e controlado de uma forma
mais efetiva e se aplica a qualquer objeto/operacao.
}
